Add New Bogosorts to Run All
Additionally,
The bogosorts are moved to after all the other sorts,
and are roughly sorted by speed.
They are all set to a delay of 1 millisecond,
and the array sizes aim for 20 visual seconds to sort.

Cocktail Bogosort's unreasonable limit is updated,
and commented code for Quick Bogosort's worst-case pivot is provided.