improve Highlights
Checking whether an array index is highlighted is now constant-time,
using an array that stores highlight counts for each index.
This is a second attempt at faster highlights.

Note that the array uses bytes instead of ints to decrease size,
so the value is capped at the maximum unsigned byte value,
at which point it will revert to scanning the highlights array.

Some other minor changes are made as well.